Fixed up how components are exported
