+++ js/src/common/helpers/index.js
@@ -1,0 +1,9 @@
+import recipientCountLabel from "./recipientCountLabel";
+import recipientLabel from "./recipientLabel";
+import recipientsLabel from "./recipientsLabel";
+
+export const helpers = {
+    'recipientCountLabel': recipientCountLabel,
+    'recipientLabel': recipientLabel,
+    'recipientsLabel': recipientsLabel,
+};
