+++ js/src/forum/index.js
@@ -19,6 +19,9 @@ import RecipientsModified from "./components/RecipientsModified";
 import RecipientLeft from './components/RecipientLeft';
 import addPrivateDiscussionSessionDropdown from './addPrivateDiscussionsToSessionDropdown';
 
+export * from '../common/helpers';
+export * from './components';
+
 app.initializers.add('fof-byobu', function (app) {
     app.routes.private_discussions = { path: '/private-discussions', component: PrivateDiscussionIndex.component() };
 
@@ -79,5 +82,3 @@ app.initializers.add('fof-byobu', function (app) {
         });
     });
 });
-
-export * from './compat';
